ci: fix valgrind UAF root cause, CodeQL overflow, Windows symlink check
The previous commit (87185ea) added wildcard valgrind suppressions for
per-route firing paths under the rationale that the read was a "false
positive." It was not. `test_utils::as_shared(stack_resource)` wrapped
stack-allocated http_resources in a shared_ptr with a no-op deleter; the
webserver kept those shared_ptrs in its route_table_, but the underlying
storage died when the test body returned. MHD's request_completed
callback fires from a daemon worker thread during `ws->stop()` (called
in tear_down, AFTER the test body's locals have already destructed), so
the per-route firing path dereferenced freed stack memory through
`mr->resource_weak_.lock()`. The "Conditional jump on uninitialised
value" valgrind report was a real UAF, not a control-block read race.

Fix: migrate every `as_shared(stack_obj)` call site (226 across 11
files) to `std::make_shared<T>(...)`. With heap-owned resources the
lifetime model becomes correct -- the webserver's shared_ptr keeps the
resource alive until `~webserver_impl` runs, which is after `stop()`
has drained MHD's workers. The `as_shared` helper is removed from
test_utils.hpp; the file now carries a note documenting why it was
deleted. The two wildcard suppressions in test/libhttpserver.supp are
also removed; only the legitimate `gnutls_session_get_data` Memcheck:Cond
suppression remains.

Also fixed on this branch:
* src/detail/ip_representation.cpp: CodeQL flagged `(16 - i) *
  a.pieces[i]` as an int*int -> int64_t overflow path. Cast the (16-i)
  factor to int64_t so the multiply is performed in the destination
  type.
* src/Makefile.am install-data-hook: `ln -s` on MSYS2/mingw silently
  falls back to a file copy (sometimes failing entirely without admin
  rights). Wrap the symlink creation in `{ ln -s ... || cp ...; }` so
  the alias is always installed under one form or another.
* Makefile.am check-install-layout: relax `test -L httpserverpp` to
  `test -e httpserverpp`. Both a POSIX symlink and an MSYS2 file copy
  satisfy the "include resolves to umbrella" property the hook is
  establishing.

The file-size ceiling temp-bump from 87185ea is left in place; walking
the nine offenders back below 500 LOC is the documented follow-up and
is independent of this fix.

+++ src/detail/ip_representation.cpp
@@ -268,8 +268,12 @@ void accumulate_octet_score(const ip_representation& a,
                             const ip_representation& b, int i,
                             int64_t& a_score, int64_t& b_score) {
     if (!(CHECK_BIT(a.mask, i) && CHECK_BIT(b.mask, i))) return;
-    a_score += (16 - i) * a.pieces[i];
-    b_score += (16 - i) * b.pieces[i];
+    // Cast the (16 - i) factor to int64_t before the multiply so the product
+    // is computed in the destination type. Without the cast the multiplication
+    // happens in int and only the result is widened, which CodeQL flags as a
+    // potential overflow.
+    a_score += static_cast<int64_t>(16 - i) * a.pieces[i];
+    b_score += static_cast<int64_t>(16 - i) * b.pieces[i];
 }
 
 // True when both v4-mapped-prefix octets on a and b are 0x00 or 0xFF.
